Add test for PackedType
(DIS-1669)

Added a test file for `PackedType` types defined in `cstruct`. For now
only the `float16` and `float` are tested, will add more over time.
